requirements.txt
ffmpeg-python==0.2.0
moviepy==1.0.3
openai==1.35.15
python-docx==1.1.2
python-dotenv==1.0.1
streamlit==1.49.0
azure-data-tables==12.5.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
bcrypt==4.2.0
httpx==0.27.0
pypdf
APScheduler
python-dateutil
openpyxl